requirements.txt
sklearn
azureml-sdk
inference-schema
python-dotenv